requirements.txt
-f /usr/share/pip-wheels
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
